% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_fit.R
\name{nb_fit}
\alias{nb_fit}
\title{Fit the negative binomial}
\usage{
nb_fit(x)
}
\arguments{
\item{x}{a site by species matrix}
}
\value{
A \code{data.frame} with columns giving the species IDs, the two
    parameters of the negative binomial distribution, the log likelihood
    of the negative binomial distribution, and the log likelihood of the
    Poisson distribution.  Each row is one species and the species ID refers
    to the column index of that species in the site by species matrix
    \code{x}.
}
\description{
Fits the negative binomial SSAD to a site x species matrix
}
\details{
If likelihood optimization fails for a given species, \code{NA}
    will be returned for all row elements corresponding to that species.
}
