% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_fit.R
\name{nb_mle}
\alias{nb_mle}
\title{Maximum likelihood estimation of negative binomial distribution}
\usage{
nb_mle(x)
}
\arguments{
\item{x}{an integer vector}
}
\value{
A \code{lsit} with components \code{estiamte} giving the MLE and
    \code{loglik} giving the value of the log likelihood function at the
    maximum. If \code{var(x) <= mean(x)} then the MLE does not exist and the
    function returns \code{size = Inf}.
}
\description{
Finds the maximum likelihood estimates of the dispersion (k)
    and mean (mu) parameters of the negative binomial distribution.
}
\details{
This function profiles the log likelihood assuming
    \code{mu = mean(x)} and uses \code{optimize} to find \code{k} that
    maximizes the log likelihood function built from \code{dnbinom}.

    Maximizing the profile likelihood is faster and avoids computational
    issues when \code{var(x)} is close to \var{mean(x)}
}
