% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_zsq.R
\name{nb_zsq}
\alias{nb_zsq}
\title{Z-squared test statistic of the negative binomial distribution}
\usage{
nb_zsq(x, size, mu)
}
\arguments{
\item{x}{a numeric vector}

\item{size}{the size parameter}

\item{mu}{the mean parameter}
}
\value{
The z-squared statistic (a numeric of length 1)
}
\description{
Exact test of the negative binomial
}
\details{
Compares the observed log likelihood of the nagative binomial distribution
given the data to the sampling distribution of log likelihoods given the negative
binomial distribution is true.  This test statistic is Chi-square distributed with
d.f. = 1, meaning value larger than \code{qchisq(0.95, 1)} rejects the "null"
hypothesis that the data come from a negative binomial.
}
