% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_stats.R
\name{net_stats}
\alias{net_stats}
\title{Make a network and calculate statistics}
\usage{
net_stats(x, alpha = 0.05, B = 999, dist_fun = "schoener")
}
\arguments{
\item{x}{site by species matrix (sites as rows, species as columns,
abundances in cells)}

\item{alpha}{the significance cut off level}

\item{B}{number of replicates for the null model permutations}

\item{dist_fun}{character name of the distance function to use
(default \code{"schoener"}). Note that the function must
calculate distances between columns, not rows}
}
\value{
a named vector of network summary statistics
}
\description{
Make a network based on a distance-like function and compute
    network summary statistics
}
