% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_nets.R
\name{sim_nets}
\alias{sim_nets}
\title{Simulate site by species matrices and compute networks from them}
\usage{
sim_nets(
  sad_stats,
  mc_cores,
  ssad_type = "nbinom",
  k_fun,
  nsim,
  alpha = 0.05,
  B = 999,
  dist_fun = "schoener"
)
}
\arguments{
\item{sad_stats}{a \code{data.frame} with columns \code{mod}, \code{par1},
\code{par2}}

\item{mc_cores}{number of cores to use in \code{parallel::mclapply}}

\item{ssad_type}{string specifying SSAD shape (e.g. \code{"nbinom"})}

\item{k_fun}{function to relate k parameter of the SSAD to relative abundance}

\item{nsim}{number of simulations to run}

\item{alpha}{significance cut off for comparing observed species similarities
to null similarities, defaults to \code{0.05}}

\item{dist_fun}{character name of the distance function to use for
calculating species similarities (default \code{"schoener"})}

\item{nsite}{number of sites to simulate}

\item{nspp}{number of species to simulate}

\item{sadfun}{function to generate random SAD sample}

\item{ssadfun}{function to generate random SSAD sample}
}
\value{
a \code{data.frame} with \code{<= nsim} rows (some simulations may be
    thrown out if they do not meet data filtering requirements), and columns
    corresponding to summary statistics about network characteristics
}
\description{
Networks are built from simulated data that mirrors real
    SADs and SSADs but has no real interactions/associations. Then summary
    statistics are calculated for those networks
}
\details{
\code{sim_nets} draws random SAD and SSAD shapes from real data
    and uses these to simulate more data and calculate network statistics on
    those simulated data. \code{simple_sim_nets} assumes one SAD and one S
    SAD and simulates data from those, again calculated network statistics.

    Note: any value passed to \code{ssad_type} other than \code{"nbinom"}
    results in a Poisson SSAD (i.e., there are only two options, negative
    binomial specified by \code{'nbinom'} or Poisson specified by anything
    else)
}
